moved config to stroke_config
