+++ __init__.py
@@ -1,10 +1,10 @@
-from . import bids_loader, bids_workflow, config, download_data, indi_reformat
+from . import bids_loader, bids_workflow, stroke_config, download_data, indi_reformat
 from . import nii_slice, prediction, problem, scoring
 
 __all__ = [
     'bids_loader',
     'bids_workflow',
-    'config',
+    'stroke_config',
     'download_data',
     'indi_reformat',
     'nii_slice',
