+++ problem.py
@@ -6,7 +6,7 @@ import warnings
 import prediction
 from bids_workflow import BIDSWorkflow
 from scoring import DiceCoeff
-import config
+import stroke_config
 from bids_loader import BIDSLoader
 
 problem_title = "ATLAS Stroke Lesion Segmentation"
@@ -37,9 +37,9 @@ def get_cv(X: np.array,
         List of train/test indices for each fold of k-fold cross-validation.
     '''
     strat = sklearn.model_selection.ShuffleSplit(
-        n_splits=config.cross_validation['n_splits'],
-        train_size=config.cross_validation['train_size'],
-        random_state=config.cross_validation['random_state'])
+        n_splits=stroke_config.cross_validation['n_splits'],
+        train_size=stroke_config.cross_validation['train_size'],
+        random_state=stroke_config.cross_validation['random_state'])
     return strat.split(X, y)
 
 
@@ -62,13 +62,13 @@ def get_train_data(path: str):
     # optimize settings.
     if(path == '.' or path == './'):
         path = 'data'
-    if(os.path.abspath(path) == os.path.abspath(config.data_path)):
-        return config.bids_loader_train.data_list, config.bids_loader_train.target_list
+    if(os.path.abspath(path) == os.path.abspath(stroke_config.data_path)):
+        return stroke_config.bids_loader_train.data_list, stroke_config.bids_loader_train.target_list
     else:
         warnings.warn(
             f'Data path differs from that in the config file; to reduce the amount of time spent loading '
             f'files, modify config.py: data_path = {path}')
-        training_dir = os.path.join(path, config.training['dir_name'])
+        training_dir = os.path.join(path, stroke_config.training['dir_name'])
         bids_loader_train = BIDSLoader(root_dir=training_dir,
                                        data_entities=[{'subject': '',
                                                        'session': '',
@@ -79,10 +79,10 @@ def get_train_data(path: str):
                                        data_derivatives_names=['ATLAS'],
                                        target_derivatives_names=['ATLAS'],
                                        label_names=['not lesion', 'lesion'],
-                                       batch_size=config.training['batch_size'])
+                                       batch_size=stroke_config.training['batch_size'])
 
-        if(config.is_quick_test):
-            bids_loader_train.data_list[:config.num_subjects_quick_test], bids_loader_train.target_list[:config.num_subjects_quick_test]
+        if(stroke_config.is_quick_test):
+            bids_loader_train.data_list[:stroke_config.num_subjects_quick_test], bids_loader_train.target_list[:stroke_config.num_subjects_quick_test]
         else:
             return bids_loader_train.data_list, bids_loader_train.target_list
 
@@ -105,13 +105,13 @@ def get_test_data(path: str):
     # optimize settings.
     if (path == '.' or path == './'):
         path = 'data'
-    if (os.path.abspath(path) == os.path.abspath(config.data_path)):
-        return config.bids_loader_test.data_list, config.bids_loader_test.target_list
+    if (os.path.abspath(path) == os.path.abspath(stroke_config.data_path)):
+        return stroke_config.bids_loader_test.data_list, stroke_config.bids_loader_test.target_list
     else:
         warnings.warn(
-            f'Data path differs from that in the config file; to reduce the amount of time spent loading '
-            f'files, modify config.py: data_path = {path}')
-        testing_dir = os.path.join(path, config.testing['dir_name'])
+            f'Data path differs from that in the stroke_config file; to reduce the amount of time spent loading '
+            f'files, modify stroke_config.py: data_path = {path}')
+        testing_dir = os.path.join(path, stroke_config.testing['dir_name'])
         bids_loader_test = BIDSLoader(root_dir=testing_dir,
                                       data_entities=[{'subject': '',
                                                       'session': '',
@@ -122,9 +122,9 @@ def get_test_data(path: str):
                                       data_derivatives_names=['ATLAS'],
                                       target_derivatives_names=['ATLAS'],
                                       label_names=['not lesion', 'lesion'],
-                                      batch_size=config.testing['batch_size'])
+                                      batch_size=stroke_config.testing['batch_size'])
 
-    if(config.is_quick_test):
-        return bids_loader_test.data_list[:config.num_subjects_quick_test], bids_loader_test.target_list[:config.num_subjects_quick_test]
+    if(stroke_config.is_quick_test):
+        return bids_loader_test.data_list[:stroke_config.num_subjects_quick_test], bids_loader_test.target_list[:stroke_config.num_subjects_quick_test]
     else:
         return bids_loader_test.data_list, bids_loader_test.target_list
