+++ stroke_config.py
